First mostly successful pass
This is the most straight forward approach. Serial execution using
callbacks all the way down. There is some ugly stuff here. The pocket results
need to be sorted inline. The spreadsheet rows are added individually and
uses an async counter. The row adds are not serialized and can potentially
be added out of order due to race conditions.