First pass at promises implementation. Not working
This is an attempt at replicating the naive approach, only using promises.
The Q library is used to wrap the methods from the api modules and promises
are returned from async calls throughout. It was much easier to deal with
calls that I wanted to happen in serial or in  parallel.

However the script isn't working. There is no output from running the script,
and the test file fails an assertion. Early attempts to debug have been
frustrating. There are no errors thrown, and I haven't found the right places
to place log lines. I'm sure this is just my own lack of familiarity though.